+++ backend/app-store-service/src/main/kotlin/app/store/rpc/ToolController.kt
@@ -4,19 +4,17 @@ import com.fasterxml.jackson.databind.JsonMappingException
 import com.fasterxml.jackson.dataformat.yaml.snakeyaml.error.MarkedYAMLException
 import com.fasterxml.jackson.module.kotlin.readValue
 import dk.sdu.cloud.CommonErrorMessage
-import dk.sdu.cloud.app.store.api.AppStore
 import dk.sdu.cloud.app.store.api.ToolDescription
 import dk.sdu.cloud.app.store.api.ToolStore
 import dk.sdu.cloud.app.store.services.LogoService
 import dk.sdu.cloud.app.store.services.LogoType
-import dk.sdu.cloud.app.store.services.ToolDAO
+import dk.sdu.cloud.app.store.services.ToolAsyncDao
 import dk.sdu.cloud.app.store.util.yamlMapper
 import dk.sdu.cloud.calls.server.HttpCall
 import dk.sdu.cloud.calls.server.RpcServer
 import dk.sdu.cloud.calls.server.securityPrincipal
 import dk.sdu.cloud.calls.types.BinaryStream
 import dk.sdu.cloud.service.Controller
-import dk.sdu.cloud.service.db.DBSessionFactory
 import dk.sdu.cloud.service.db.async.AsyncDBSessionFactory
 import dk.sdu.cloud.service.db.withTransaction
 import io.ktor.application.call
@@ -30,7 +28,7 @@ import java.io.ByteArrayInputStream
 
 class ToolController(
     private val db: AsyncDBSessionFactory,
-    private val toolDao: ToolDAO,
+    private val toolDao: ToolAsyncDao,
     private val logoService: LogoService
 ) : Controller {
     override fun configure(rpcServer: RpcServer) = with(rpcServer) {
