+++ backend/app-store-service/src/main/kotlin/app/store/services/ApplicationSearchService.kt
@@ -1,21 +1,14 @@
-package app.store.services
+package dk.sdu.cloud.app.store.services
 
 import com.fasterxml.jackson.module.kotlin.readValue
-import com.github.jasync.sql.db.RowData
 import dk.sdu.cloud.SecurityPrincipal
+import dk.sdu.cloud.app.store.api.Application
 import dk.sdu.cloud.app.store.api.ApplicationSummaryWithFavorite
-import dk.sdu.cloud.app.store.services.AppStoreAsyncDAO
-import dk.sdu.cloud.app.store.services.ApplicationTable
-import dk.sdu.cloud.app.store.services.ElasticDAO
-import dk.sdu.cloud.app.store.services.ElasticIndexedApplication
-import dk.sdu.cloud.app.store.services.EmbeddedNameAndVersion
-import dk.sdu.cloud.app.store.services.toApplicationWithInvocation
 import dk.sdu.cloud.calls.client.AuthenticatedClient
 import dk.sdu.cloud.defaultMapper
 import dk.sdu.cloud.service.NormalizedPaginationRequest
 import dk.sdu.cloud.service.Page
 import dk.sdu.cloud.service.db.async.AsyncDBSessionFactory
-import dk.sdu.cloud.service.db.async.getField
 import dk.sdu.cloud.service.db.async.withSession
 import dk.sdu.cloud.service.mapItems
 import dk.sdu.cloud.service.paginate
@@ -23,9 +16,9 @@ import org.elasticsearch.action.search.SearchResponse
 
 class ApplicationSearchService (
     private val db: AsyncDBSessionFactory,
-    private val searchDAO: ApplicationSearchAsyncDAO,
-    private val elasticDAO: ElasticDAO,
-    private val applicationDAO: AppStoreAsyncDAO,
+    private val searchDao: ApplicationSearchAsyncDao,
+    private val elasticDao: ElasticDao,
+    private val applicationDao: AppStoreAsyncDao,
     private val authenticatedClient: AuthenticatedClient
 ) {
     suspend fun searchByTags(
@@ -41,7 +34,7 @@ class ApplicationSearchService (
         }
 
         return db.withSession { session ->
-            searchDAO.searchByTags(
+            searchDao.searchByTags(
                 session,
                 securityPrincipal,
                 project,
@@ -65,7 +58,7 @@ class ApplicationSearchService (
         }
 
         return db.withSession { session ->
-            searchDAO.search(
+            searchDao.search(
                 session,
                 securityPrincipal,
                 project,
@@ -107,7 +100,7 @@ class ApplicationSearchService (
 
         val queryTerms = normalizedQuery.split(" ").filter { it.isNotBlank() }
 
-        val results = elasticDAO.search(queryTerms, normalizedTags)
+        val results = elasticDao.search(queryTerms, normalizedTags)
 
         if (results.hits.hits.isEmpty()) {
             return Page(
@@ -131,7 +124,7 @@ class ApplicationSearchService (
             }
 
             val applications = db.withSession { session ->
-                applicationDAO.findAllByID(session, user, project, projectGroups, embeddedNameAndVersionList, paging)
+                applicationDao.findAllByID(session, user, project, projectGroups, embeddedNameAndVersionList, paging)
             }
 
             return sortAndCreatePageByScore(applications, results, user, paging)
@@ -143,7 +136,7 @@ class ApplicationSearchService (
             }
 
             val applications = db.withSession { session ->
-                searchDAO.multiKeywordsearch(session, user, project, projectGroups, titles.toList(), paging)
+                searchDao.multiKeywordsearch(session, user, project, projectGroups, titles.toList(), paging)
             }
 
             return sortAndCreatePageByScore(applications, results, user, paging)
@@ -151,19 +144,19 @@ class ApplicationSearchService (
     }
 
     private suspend fun sortAndCreatePageByScore(
-        applications: List<RowData>,
+        applications: List<Application>,
         results: SearchResponse,
         user: SecurityPrincipal,
         paging: NormalizedPaginationRequest
     ): Page<ApplicationSummaryWithFavorite> {
         val map = applications.associateBy(
             { EmbeddedNameAndVersion(
-                it.getField(ApplicationTable.idName).toLowerCase(),
-                it.getField(ApplicationTable.idVersion).toLowerCase()
+                it.metadata.name.toLowerCase(),
+                it.metadata.version.toLowerCase()
             ) }, { it }
         )
 
-        val sortedList = mutableListOf<RowData>()
+        val sortedList = mutableListOf<Application>()
 
         results.hits.hits.forEach {
             val foundEntity =
@@ -173,10 +166,10 @@ class ApplicationSearchService (
             }
         }
 
-        val sortedResultsPage = sortedList.map { it.toApplicationWithInvocation() }.paginate(paging)
+        val sortedResultsPage = sortedList.paginate(paging)
 
         return db.withSession { session ->
-            applicationDAO.preparePageForUser(session, user.username, sortedResultsPage)
+            applicationDao.preparePageForUser(session, user.username, sortedResultsPage)
                 .mapItems { it.withoutInvocation() }
         }
     }
