+++ backend/app-store-service/src/test/kotlin/app/store/services/ElasticDaoTest.kt
@@ -4,32 +4,23 @@ import dk.sdu.cloud.app.store.util.normAppDesc
 import dk.sdu.cloud.calls.RPCException
 import io.mockk.every
 import io.mockk.mockk
-import org.apache.lucene.search.TotalHits
-import org.elasticsearch.action.OriginalIndices
 import org.elasticsearch.action.admin.indices.flush.FlushResponse
 import org.elasticsearch.action.search.SearchResponse
 import org.elasticsearch.action.update.UpdateResponse
 import org.elasticsearch.client.IndicesClient
 import org.elasticsearch.client.RestHighLevelClient
 import org.elasticsearch.client.indices.GetIndexRequest
-import org.elasticsearch.common.bytes.BytesArray
-import org.elasticsearch.common.document.DocumentField
-import org.elasticsearch.common.text.Text
-import org.elasticsearch.index.Index
-import org.elasticsearch.index.shard.ShardId
 import org.elasticsearch.search.SearchHit
 import org.elasticsearch.search.SearchHits
-import org.elasticsearch.search.SearchShardTarget
-import org.elasticsearch.search.internal.InternalSearchResponse
 import org.junit.Test
 import kotlin.test.assertEquals
 
-class ElasticDAOTest {
+class ElasticDaoTest {
 
     @Test
     fun testSearch() {
         val elasticHighLevelClient = mockk<RestHighLevelClient>()
-        val elasticDAO = ElasticDAO(elasticHighLevelClient)
+        val elasticDAO = ElasticDao(elasticHighLevelClient)
 
         every { elasticHighLevelClient.search(any(), any()) } answers {
             val response = mockk<SearchResponse>()
@@ -53,7 +44,7 @@ class ElasticDAOTest {
     @Test
     fun `Search test with title`() {
         val elasticHighLevelClient = mockk<RestHighLevelClient>()
-        val elasticDAO = ElasticDAO(elasticHighLevelClient)
+        val elasticDAO = ElasticDao(elasticHighLevelClient)
 
         every { elasticHighLevelClient.search(any(), any()) } answers {
             val response = mockk<SearchResponse>()
@@ -77,7 +68,7 @@ class ElasticDAOTest {
     @Test (expected = RPCException::class)
     fun `Create test - multiple entries`() {
         val elasticHighLevelClient = mockk<RestHighLevelClient>()
-        val elasticDAO = ElasticDAO(elasticHighLevelClient)
+        val elasticDAO = ElasticDao(elasticHighLevelClient)
 
         every { elasticHighLevelClient.indices() } answers {
             val client = mockk<IndicesClient>()
@@ -105,7 +96,7 @@ class ElasticDAOTest {
     @Test
     fun `Create test`() {
         val elasticHighLevelClient = mockk<RestHighLevelClient>()
-        val elasticDAO = ElasticDAO(elasticHighLevelClient)
+        val elasticDAO = ElasticDao(elasticHighLevelClient)
 
         every { elasticHighLevelClient.indices() } answers {
             val client = mockk<IndicesClient>()
@@ -133,7 +124,7 @@ class ElasticDAOTest {
     @Test
     fun `Create No index test`() {
         val elasticHighLevelClient = mockk<RestHighLevelClient>(relaxed = true)
-        val elasticDAO = ElasticDAO(elasticHighLevelClient)
+        val elasticDAO = ElasticDao(elasticHighLevelClient)
 
         elasticDAO.createApplicationInElastic(normAppDesc)
     }
@@ -141,7 +132,7 @@ class ElasticDAOTest {
     @Test
     fun `Update Description test`() {
         val elasticHighLevelClient = mockk<RestHighLevelClient>()
-        val elasticDAO = ElasticDAO(elasticHighLevelClient)
+        val elasticDAO = ElasticDao(elasticHighLevelClient)
 
         every { elasticHighLevelClient.indices() } answers {
             val client = mockk<IndicesClient>()
@@ -183,7 +174,7 @@ class ElasticDAOTest {
     @Test (expected = RPCException::class)
     fun `Update Description test - not found`() {
         val elasticHighLevelClient = mockk<RestHighLevelClient>()
-        val elasticDAO = ElasticDAO(elasticHighLevelClient)
+        val elasticDAO = ElasticDao(elasticHighLevelClient)
 
         every { elasticHighLevelClient.indices() } answers {
             val client = mockk<IndicesClient>()
@@ -213,7 +204,7 @@ class ElasticDAOTest {
     @Test (expected = RPCException::class)
     fun `Update Description test - multiple entries`() {
         val elasticHighLevelClient = mockk<RestHighLevelClient>()
-        val elasticDAO = ElasticDAO(elasticHighLevelClient)
+        val elasticDAO = ElasticDao(elasticHighLevelClient)
 
         every { elasticHighLevelClient.indices() } answers {
             val client = mockk<IndicesClient>()
